api/pyproject.toml
[project]
name = "forex-api"
version = "0.1.0"
description = "FastAPI service for forex TWI data."
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "google-cloud-bigquery>=3.26.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pyyaml>=6.0.0",
    "ruff>=0.6.9",
    "ty>=0.0.1",
]
